Yeah

+++ No bonus/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: randre <randre@student.s19.be>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/23 06:56:59 by randre            #+#    #+#             */
-/*   Updated: 2023/10/07 14:37:42 by randre           ###   ########.fr       */
+/*   Updated: 2023/10/07 14:47:40 by randre           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ int				ft_typecheck(const char c, va_list args);
 int				ft_putnbr(int n);
 char			*ft_strdup(const char *s);
 int				ft_xhandle(va_list args, size_t i);
-size_t			ft_xcount(size_t nbr);
+int				ft_xcount(size_t nbr);
 int				ft_putnbr_long(long long int n);
 int				ft_hexa(size_t nbr);
 int				ft_phandle(va_list args);
